jx3_script/requirements.txt
pytesseract~=0.3.13
opencv-python~=4.10.0.84
numpy~=2.1.1
pillow~=10.4.0
pywin32~=306